requirements.txt
flask
numpy
tensorflow
Pillow
opencv-python-headless
scipy